upgrade httpsnippet dependency
I was noticing errors in some of our code snippets that rely on
openapi-snippet. In particular string parameters with double quotes in them
would not have the quotes properly escaped in C# examples.

This commit updates the version of httpsnippet we use which I noticed
had commits that fixed problems like this.

I then added a new test to ensure quotes are escaped properly in CSharp strings.

I assume we'll get plenty of other language fixes as well.

I also had to remove use of `? :` and `??` operators since some dependency did not
like using the version of node that supports those operators.